Quiz Link -> Profile, Housing Search
Added a housing page search bar. Added a button to link the similar users profiles to their other user profile page. Updated email href in other users profiles so that you can contact them. Added it so that you can't do AND see the quiz unless you are logged in.